feldy模块，移除dispaly、readonly之类的属性
