+++ examples/normal-editor-example/src/data/details.json
@@ -63,7 +63,6 @@
               "componentName": "Input",
               "props": {},
               "x-field": {
-                "reactionMeta": {},
                 "name": "test1"
               },
               "slots": {},
@@ -73,15 +72,6 @@
               "componentName": "Input",
               "props": {},
               "x-field": {
-                "reactionMeta": {
-                  "display": {},
-                  "pattern": {},
-                  "hidden": {
-                    "expression": "$test1.value === \"1\""
-                  },
-                  "disabled": {},
-                  "readonly": {}
-                },
                 "name": "test2"
               },
               "slots": {},
