+++ examples/normal-editor-example/src/data/list.json
@@ -104,15 +104,7 @@
                               "componentName": "Input",
                               "props": {},
                               "x-field": {
-                                "name": "",
-                                "params": {},
-                                "reactionMeta": {
-                                  "display": {},
-                                  "pattern": {},
-                                  "hidden": {},
-                                  "disabled": {},
-                                  "readonly": {}
-                                }
+                                "name": ""
                               },
                               "x-controller": {
                                 "id": "67ba7358-b919-4234-8ae9-b0a0d666d3c3",
@@ -1278,15 +1270,6 @@
                           }
                         ]
                       },
-                      "x-field": {
-                        "reactionMeta": {
-                          "display": {},
-                          "pattern": {},
-                          "hidden": {},
-                          "disabled": {},
-                          "readonly": {}
-                        }
-                      },
                       "bordered": false,
                       "slots": {},
                       "children": [
@@ -1296,14 +1279,7 @@
                             "title": "名称"
                           },
                           "x-field": {
-                            "name": "name",
-                            "reactionMeta": {
-                              "display": {},
-                              "pattern": {},
-                              "hidden": {},
-                              "disabled": {},
-                              "readonly": {}
-                            }
+                            "name": "name"
                           },
                           "slots": {},
                           "children": []
@@ -1314,14 +1290,7 @@
                             "title": "登录名称"
                           },
                           "x-field": {
-                            "name": "login_name",
-                            "reactionMeta": {
-                              "display": {},
-                              "pattern": {},
-                              "hidden": {},
-                              "disabled": {},
-                              "readonly": {}
-                            }
+                            "name": "login_name"
                           },
                           "slots": {},
                           "children": []
@@ -1365,29 +1334,13 @@
                                           },
                                           "locked": false,
                                           "x-field": {
-                                            "name": "login_name",
-                                            "reactionMeta": {
-                                              "display": {},
-                                              "pattern": {},
-                                              "hidden": {},
-                                              "disabled": {},
-                                              "readonly": {}
-                                            }
+                                            "name": "login_name"
                                           },
                                           "slots": {},
                                           "children": [
                                             {
                                               "componentName": "Input",
                                               "props": {},
-                                              "x-field": {
-                                                "reactionMeta": {
-                                                  "display": {},
-                                                  "pattern": {},
-                                                  "hidden": {},
-                                                  "disabled": {},
-                                                  "readonly": {}
-                                                }
-                                              },
                                               "slots": {},
                                               "children": []
                                             }
@@ -1400,29 +1353,13 @@
                                           },
                                           "locked": false,
                                           "x-field": {
-                                            "name": "name",
-                                            "reactionMeta": {
-                                              "display": {},
-                                              "pattern": {},
-                                              "hidden": {},
-                                              "disabled": {},
-                                              "readonly": {}
-                                            }
+                                            "name": "name"
                                           },
                                           "slots": {},
                                           "children": [
                                             {
                                               "componentName": "Input",
                                               "props": {},
-                                              "x-field": {
-                                                "reactionMeta": {
-                                                  "display": {},
-                                                  "pattern": {},
-                                                  "hidden": {},
-                                                  "disabled": {},
-                                                  "readonly": {}
-                                                }
-                                              },
                                               "slots": {},
                                               "children": []
                                             }
