+++ packages/fieldy/fieldy-react/src/components/XField/index.tsx
@@ -1,8 +1,7 @@
-import { DisplayType, IFieldMeta } from "@rxdrag/fieldy"
-import React, { memo, useCallback, useEffect, useState } from "react"
+import { IFieldMeta } from "@rxdrag/fieldy"
+import React, { memo, useEffect } from "react"
 import { FieldContext } from "../../contexts"
 import { useRegisterField } from "./hooks/useRegisterField"
-import { FieldState } from "@rxdrag/fieldy"
 
 export const XField = memo((props: {
   fieldMeta: IFieldMeta,
@@ -12,11 +11,7 @@ export const XField = memo((props: {
   defaultValue?: unknown,
 }) => {
   const { fieldMeta, initialValue, value, defaultValue, children } = props
-  const [hidden, setHidden] = useState(false);
   const field = useRegisterField(fieldMeta)
-  const handleFieldChange = useCallback((fieldState: FieldState | undefined) => {
-    setHidden(fieldState?.hidden || fieldState?.display === DisplayType.hidden || fieldState?.display === DisplayType.none)
-  }, [])
 
   useEffect(() => {
     initialValue !== undefined && field?.setInitialValue(initialValue)
@@ -30,19 +25,10 @@ export const XField = memo((props: {
     defaultValue !== undefined && field?.setDefaultValue(defaultValue)
   }, [field, defaultValue])
 
-  useEffect(() => {
-    setHidden(field?.form.getFieldState(field.path)?.hidden || false)
-  }, [field?.form, field?.path])
-
-  useEffect(() => {
-    const unsub = field?.fieldy.subscribeToFieldChange(field.form.name, field.path, handleFieldChange)
-    return unsub
-  }, [field?.fieldy, field?.form.name, field?.path, handleFieldChange])
-
   return (
     <FieldContext.Provider value={field}>
       {
-        hidden === false && children
+        children
       }
     </FieldContext.Provider>
   )
