+++ packages/fieldy/fieldy/src/reducers/forms/form/fields.ts
@@ -1,10 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable no-case-declarations */
-import { isStr } from "@rxdrag/shared";
 import { SET_FORM_FIELDS, SetFormFieldsPayload, ADD_FORM_FIELDS, REMOVE_FORM_FIELD, RemoveFormFieldPayload, FieldActionPayload, SetFieldValuePayload, SET_FORM_INITIAL_VALUE, SET_FORM_FIELDS_FEEDBACKS, SetFormFeedbacksPayload } from "../../../actions";
 import { FieldsState, IAction, IFieldSchema } from "../../../interfaces/fieldy";
 import { fieldReduce } from "./field";
-import { DisplayType, Expression, PatternType } from "../../../interfaces";
 
 let idSeed = 1
 function makeId() {
@@ -81,39 +79,10 @@ function makeFields(fieldSchemas: IFieldSchema[]) {
         basePath: schema.path.substring(0, schema.path.length - (schema.name?.toString().length || 0) - 1),
         mounted: true,
         meta: schema,
-        display: extractValue(schema?.reactionMeta?.display) as DisplayType | undefined,
-        pattern: extractValue(schema?.reactionMeta?.pattern) as PatternType | undefined,
-        hidden: extractValue(schema?.reactionMeta?.hidden) as boolean | undefined,
-        disabled: extractValue(schema?.reactionMeta?.disabled) as boolean | undefined,
-        readonly: extractValue(schema?.reactionMeta?.readonly) as boolean | undefined,
       }
     }
 
     return flatFields
   }
 
-  type ExpressionObj = {
-    value?: unknown,
-    expression?: string,
-  }
-
-  function extractValue(expOrValue?: {
-    value?: unknown,
-    expression?: string,
-  } | Expression) {
-    if (isStr(expOrValue)) {
-      if (expOrValue.trim().startsWith("{{")) {
-        return undefined
-      }
-      return expOrValue;
-    }
-    if (expOrValue?.expression) {
-      return expOrValue.value
-    }
-    const keys = Object.keys(expOrValue || {})
-    if (keys.find((key) => key === "value") || keys.find((key) => key === "expression")) {
-      (expOrValue as ExpressionObj)?.value
-    }
-    return keys.length == 0 ? undefined : expOrValue
-  }
 }
