+++ packages/rxdrag/editor/react-antd-materials/src/shared/createSchema.ts
@@ -61,9 +61,6 @@ export function createSchema(options: SchemaOptions = {}): INodeSchema {
       props: {
         title: "$field",
         id: "data",
-        style: {
-          padding: 0,
-        }
       },
       children: createFieldSchema()
     }] : []
