+++ packages/rxdrag/editor/react-antd-shell/src/locales.ts
@@ -38,7 +38,7 @@ export const commonLocales: ILocales = {
       design: "设计",
       jsonCode: "JSON 代码",
       preview: "预览",
-      addView: "添加视图",
+      views: "视图管理",
       pointSelection: "点选",
       boxSelection: "框选",
       reset: "重置"
@@ -78,7 +78,7 @@ export const commonLocales: ILocales = {
       design: "Design",
       jsonCode: "JSON Code",
       preview: "Preview",
-      addView: "Add View",
+      views: "Views",
       pointSelection: "Point selection",
       boxSelection: "Box selection",
       reset: "Reset",
