+++ packages/rxdrag/runtime/react-antd-components/src/forms/index.ts
@@ -2,3 +2,4 @@ export * from "./Form"
 export * from "./InlineForm"
 export * from "./FormItem"
 export * from "./FormLayout"
+export * from "./types"
