+++ packages/rxdrag/runtime/react-antd-components/src/forms/types.ts
@@ -1,0 +1,19 @@
+export enum PatternType {
+  //可编辑
+  editable = 'editable',
+  //禁用
+  disabled = 'disabled',
+  //只读
+  readOnly = 'readOnly',
+  //阅读模式，比如把input转成text，需要组件支持
+  readPretty = 'readPretty'
+}
+
+export enum DisplayType {
+  //代表字段 UI 隐藏，同时不保留字段数据
+  none = "none",
+  //代表字段 UI 隐藏，保留字段数据
+  hidden = "hidden",
+  //代表字段 UI 显示，同时恢复字段数据
+  visible = "visible"
+}
